requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==5.0.2
pydantic==2.6.1
pydantic_core==2.16.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
psycopg2-binary
django-ninja
django-ninja-extra
python-jose

